Adjusted the interfaces and base classes to properly allow for non-generic patterns. #2
