+++ src/Foxlabs.Domain.Abstractions/AggregateRoot.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text.Json.Serialization;
 
 namespace FoxLabs.Domain
 {
@@ -33,8 +32,6 @@ namespace FoxLabs.Domain
         where TAggregate : class, IAggregateRoot<TKey>
         where TKey : IComparable
     {
-        private List<IDomainEvent<TKey>> _domainEvents;
-
         protected AggregateRoot()
             : base() { }
 
@@ -43,35 +40,9 @@ namespace FoxLabs.Domain
         { }
 
         /// <summary>
-        /// The read-only collection of domain events for the entity.
-        /// </summary>
-        [JsonIgnore]
-        public IReadOnlyCollection<IDomainEvent<TKey>> DomainEvents => _domainEvents;
-
-        /// <summary>
         /// The version of the aggregate.
         /// </summary>
         public long Version { get; protected set; }
-
-        /// <summary>
-        /// Add an <see cref="IDomainEvent" /> to the entity.
-        /// </summary>
-        public void AddDomainEvent(IDomainEvent<TKey> @event)
-        {
-            (_domainEvents ??= new List<IDomainEvent<TKey>>()).Add(@event);
-        }
-
-        /// <summary>
-        /// Clears the collection of <see cref="IDomainEvent" />s on the entity.
-        /// </summary>
-        public void ClearDomainEvents()
-            => _domainEvents?.Clear();
-
-        /// <summary>
-        /// Remove an <see cref="IDomainEvent" /> from the entity.
-        /// </summary>
-        public void RemoveDomainEvent(IDomainEvent<TKey> @event)
-            => _domainEvents?.Remove(@event);
 
         /// <summary>
         /// Create an instance of <paramref name="TAggregate" />.
