+++ src/Foxlabs.Domain.Abstractions/Entity.cs
@@ -1,11 +1,13 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Text.Json.Serialization;
 
 namespace FoxLabs.Domain
 {
     /// <summary>
     /// The base class for entities which belong to an <see cref="AggregateRoot" />.
     /// </summary>
-    public abstract class Entity: Entity<int>
+    public abstract class Entity : Entity<int>
     {
         protected Entity()
             : base() { }
@@ -21,6 +23,8 @@ namespace FoxLabs.Domain
     public abstract class Entity<TKey> : IEntity<TKey>
         where TKey : IComparable
     {
+        private List<IDomainEvent<TKey>> _domainEvents;
+
         private int? _requestedHashCode;
 
         protected Entity() { }
@@ -29,6 +33,12 @@ namespace FoxLabs.Domain
         {
             Id = id;
         }
+
+        /// <summary>
+        /// The read-only collection of domain events for the entity.
+        /// </summary>
+        [JsonIgnore]
+        public IReadOnlyCollection<IDomainEvent<TKey>> DomainEvents => _domainEvents;
 
         /// <summary>
         /// The unique identifier of this entity.
@@ -42,6 +52,30 @@ namespace FoxLabs.Domain
         /// <c>True</c> if the entity is not-persisted, otherwise <c>false</c>.
         /// </value>
         public bool IsTransient => Id?.Equals(default) ?? true;
+
+        IReadOnlyCollection<IDomainEvent> IEntity.DomainEvents
+            => DomainEvents;
+
+        /// <summary>
+        /// Add an <see cref="IDomainEvent" /> to the entity.
+        /// </summary>
+        public void AddDomainEvent(IDomainEvent<TKey> @event)
+        {
+            (_domainEvents ??= new List<IDomainEvent<TKey>>()).Add(@event);
+        }
+
+        /// <summary>
+        /// Clears the collection of <see cref="IDomainEvent" />s on the entity.
+        /// </summary>
+        public void ClearDomainEvents()
+            => _domainEvents?.Clear();
+
+        /// <summary>
+        /// Remove an <see cref="IDomainEvent" /> from the entity.
+        /// </summary>
+        public void RemoveDomainEvent(IDomainEvent<TKey> @event)
+            => _domainEvents?.Remove(@event);
+
 
         /// <inheritdoc />
         public override bool Equals(object obj)
