+++ src/Foxlabs.Domain.Abstractions/IDomainEvent.cs
@@ -6,16 +6,22 @@ namespace FoxLabs.Domain
     /// <summary>
     /// The basic interface for a domain event.
     /// </summary>
-    /// <typeparam name="TKey">The entity key type.</typeparam>
-    public interface IDomainEvent<TKey> : INotification
-        where TKey : IComparable
+    public interface IDomainEvent : INotification
     {
         IAggregateRoot RootAggregate { get; }
-
-        TKey RootId { get; }
 
         long RootVersion { get; }
 
         DateTimeOffset Timestamp { get; }
+    }
+
+    /// <summary>
+    /// The basic interface for a domain event.
+    /// </summary>
+    /// <typeparam name="TKey">The entity key type.</typeparam>
+    public interface IDomainEvent<TKey> : IDomainEvent
+        where TKey : IComparable
+    {
+        TKey RootId { get; }
     }
 }
