+++ src/Foxlabs.Domain.Abstractions/IEntity.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
 namespace FoxLabs.Domain
@@ -6,8 +7,16 @@ namespace FoxLabs.Domain
     /// <summary>
     /// The basic interface for an entity.
     /// </summary>
+    public interface IEntity
+    {
+        IReadOnlyCollection<IDomainEvent> DomainEvents { get; }
+    }
+
+    /// <summary>
+    /// The basic interface for an entity.
+    /// </summary>
     /// <typeparam name="TKey">The entity key type.</typeparam>
-    public interface IEntity<TKey>
+    public interface IEntity<TKey> : IEntity
         where TKey : IComparable
     {
         /// <summary>
@@ -20,5 +29,25 @@ namespace FoxLabs.Domain
         /// </summary>
         [JsonIgnore]
         bool IsTransient { get; }
+
+        /// <summary>
+        /// A collection of <see cref="IDomainEvent" />s.
+        /// </summary>
+        new IReadOnlyCollection<IDomainEvent<TKey>> DomainEvents { get; }
+
+        /// <summary>
+        /// Adds an <see cref="IDomainEvent" /> to the entity.
+        /// </summary>
+        void AddDomainEvent(IDomainEvent<TKey> @event);
+
+        /// <summary>
+        /// Clears the collection of <see cref="IDomainEvent" />s on the entity.
+        /// </summary>
+        void ClearDomainEvents();
+
+        /// <summary>
+        /// Remove an <see cref="IDomainEvent" /> from the entity.
+        /// </summary>
+        void RemoveDomainEvent(IDomainEvent<TKey> @event);
     }
 }
